Jetson Software/V1/requirements.txt
ttkthemes~=3.2.2
numpy~=1.21.5
pyserial~=3.5
imutils~=0.5.4
opencv-python~=4.5.5.64
nest-asyncio~=1.5.4
Pillow~=9.0.1
retry~=0.9.2
Jetson.GPIO~=2.1.0